fix: rename back + run format
