+++ www/src/LandingPageComponents/Navbar/mobileMenuButton.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource react */
 import { useStore } from "@nanostores/react";
-import { MobileMenuState } from "../../stores/MobileMenu";
+import { MobileMenuState } from "../../stores/mobileMenu";
 
 const MobileMenuButton = () => {
   const $isMobileMenuOpen = useStore(MobileMenuState);
